Presiserte modulnavn. Flyttet cache til egen modul. Fjernet test-modul. Fjernet masking package fra metrics modulen. Fjernet packages for database og kafka fra utils.
