+++ dittnav-common-evicting-cache/build.gradle.kts
@@ -10,9 +10,6 @@ dependencies {
     implementation(Logstash.logbackEncoder)
     implementation(Kotlinx.coroutines)
     testImplementation(kotlin("test-junit5"))
-    testImplementation(Jjwt.api)
-    testImplementation(Jjwt.impl)
-    testImplementation(Jjwt.jackson)
     testImplementation(Junit.engine)
     testImplementation(Kluent.kluent)
     testImplementation(Mockk.mockk)
