+++ dittnav-common-evicting-cache/src/main/kotlin/no/nav/personbruker/dittnav/common/cache/EvictingCacheConfig.kt
@@ -1,0 +1,7 @@
+package no.nav.personbruker.dittnav.common.cache
+
+data class EvictingCacheConfig (
+    val evictionThreshold: Int = 1024,
+    val entryLifetimeMinutes: Long = 15,
+    val massEvictionCoolDownMinutes: Long = 5
+)
