+++ dittnav-common-influx/build.gradle.kts
