+++ dittnav-common-influx/src/main/kotlin/no/nav/personbruker/dittnav/common/metrics/influx/SensuClient.kt
