+++ dittnav-common-influx/src/test/kotlin/no/nav/personbruker/dittnav/common/metrics/influx/UnbufferedDataPointRelayTest.kt
