--- dittnav-common-metrics/src/main/kotlin/no/nav/personbruker/dittnav/common/metrics/masking/NameAndPublicAlias.kt
@@ -1,6 +1,0 @@
-package no.nav.personbruker.dittnav.common.metrics.masking
-
-data class NameAndPublicAlias (
-    val name: String,
-    val publicAlias: String
-)
