+++ dittnav-common-security-authenticated-user/build.gradle.kts
