+++ dittnav-common-security-authenticated-user/src/main/kotlin/no/nav/personbruker/dittnav/common/security/AuthenticatedUserFactory.kt
