+++ dittnav-common-security-authenticated-user/src/test/kotlin/no/nav/personbruker/dittnav/common/security/AuthenticatedUserTest.kt
