--- dittnav-common-test/src/main/kotlin/no/nav/personbruker/dittnav/common/test/kluentExtentions.kt
@@ -1,6 +1,0 @@
-package no.nav.personbruker.dittnav.common.test
-
-import org.amshove.kluent.ExceptionResult
-import org.amshove.kluent.shouldContain
-
-infix fun <T : Throwable> ExceptionResult<T>.`with message containing`(theMessage: String) = this.exceptionMessage shouldContain (theMessage)
