--- dittnav-common-utils/src/main/kotlin/no/nav/personbruker/dittnav/common/util/cache/EvictingCacheConfig.kt
@@ -1,7 +1,0 @@
-package no.nav.personbruker.dittnav.common.util.cache
-
-data class EvictingCacheConfig (
-    val evictionThreshold: Int = 1024,
-    val entryLifetimeMinutes: Long = 15,
-    val massEvictionCoolDownMinutes: Long = 5
-)
