--- dittnav-common-utils/src/main/kotlin/no/nav/personbruker/dittnav/common/util/database/exception/BatchUpdateException.kt
@@ -1,5 +1,0 @@
-package no.nav.personbruker.dittnav.common.util.database.exception
-
-class BatchUpdateException(message: String, cause: Throwable?) : RetriableDatabaseException(message, cause) {
-    constructor(message: String) : this(message, null)
-}
