--- dittnav-common-utils/src/main/kotlin/no/nav/personbruker/dittnav/common/util/database/exception/UnretriableDatabaseException.kt
@@ -1,7 +1,0 @@
-package no.nav.personbruker.dittnav.common.util.database.exception
-
-import no.nav.personbruker.dittnav.common.util.exception.ExceptionWithContext
-
-class UnretriableDatabaseException(message: String, cause: Throwable?) : ExceptionWithContext(message, cause) {
-    constructor(message: String) : this(message, null)
-}
