--- dittnav-common-utils/src/main/kotlin/no/nav/personbruker/dittnav/common/util/database/persisting/PersistFailureReason.kt
@@ -1,5 +1,0 @@
-package no.nav.personbruker.dittnav.common.util.database.persisting
-
-enum class PersistFailureReason {
-    NO_ERROR, CONFLICTING_KEYS, UNKNOWN
-}
