--- dittnav-common-utils/src/main/kotlin/no/nav/personbruker/dittnav/common/util/database/updating/updateActionExtensions.kt
@@ -1,15 +1,0 @@
-package no.nav.personbruker.dittnav.common.util.database.updating
-
-import java.sql.Connection
-import java.sql.PreparedStatement
-
-fun Connection.executeBatchUpdateQuery(sql: String, paramInit: PreparedStatement.() -> Unit) {
-    autoCommit = false
-
-    prepareStatement(sql).use { statement ->
-        statement.paramInit()
-        statement.executeBatch()
-    }
-
-    commit()
-}
