--- dittnav-common-utils/src/main/kotlin/no/nav/personbruker/dittnav/common/util/exception/ExceptionWithContext.kt
@@ -1,18 +1,0 @@
-package no.nav.personbruker.dittnav.common.util.exception
-
-open class ExceptionWithContext(message: String, cause: Throwable?) : Exception(message, cause) {
-
-    val context: MutableMap<String, Any> = mutableMapOf()
-
-    fun addContext(key: String, value: Any) {
-        context[key] = value
-    }
-
-    override fun toString(): String {
-        return when (context.isNotEmpty()) {
-            true -> super.toString() + ", context: $context"
-            false -> super.toString()
-        }
-    }
-
-}
