--- dittnav-common-utils/src/main/kotlin/no/nav/personbruker/dittnav/common/util/kafka/RecordKeyValueWrapper.kt
@@ -1,6 +1,0 @@
-package no.nav.personbruker.dittnav.common.util.kafka
-
-data class RecordKeyValueWrapper <K, V> (
-    val key: K,
-    val value: V
-)
