--- dittnav-common-utils/src/main/kotlin/no/nav/personbruker/dittnav/common/util/kafka/exception/RetriableKafkaException.kt
@@ -1,7 +1,0 @@
-package no.nav.personbruker.dittnav.common.util.kafka.exception
-
-import no.nav.personbruker.dittnav.common.util.exception.ExceptionWithContext
-
-class RetriableKafkaException(message: String, cause: Throwable?) : ExceptionWithContext(message, cause) {
-    constructor(message: String) : this(message, null)
-}
