--- dittnav-common-utils/src/test/kotlin/no/nav/personbruker/dittnav/common/util/database/persisting/ListPersistActionResultTest.kt
@@ -1,38 +1,0 @@
-package no.nav.personbruker.dittnav.common.util.database.persisting
-
-import no.nav.personbruker.dittnav.common.util.database.persisting.PersistFailureReason.*
-import org.amshove.kluent.`should be equal to`
-import org.junit.jupiter.api.Test
-
-
-internal class ListPersistActionResultTest {
-
-    @Test
-    fun `Function mapParamListToResultArray should map persist result correctly`() {
-        val toPersist = listOf(1, 2, 3, 4)
-
-        val resultArray = listOf(1, 1, 0, 1).toIntArray()
-
-        val result = ListPersistActionResult.mapParamListToResultArray(toPersist, resultArray)
-
-        result.allEntitiesPersisted() `should be equal to` false
-        result.getPersistedEntitites() `should be equal to` listOf(1, 2, 4)
-        result.getConflictingEntities() `should be equal to` listOf(3)
-    }
-
-    @Test
-    fun `Function mapListOfIndividualResults should map persist result correctly`() {
-        val resultList = listOf(
-            1 to NO_ERROR,
-            2 to UNKNOWN,
-            3 to CONFLICTING_KEYS,
-            4 to NO_ERROR
-        )
-
-        val result = ListPersistActionResult.mapListOfIndividualResults(resultList)
-
-        result.allEntitiesPersisted() `should be equal to` false
-        result.getPersistedEntitites() `should be equal to` listOf(1, 4)
-        result.getConflictingEntities() `should be equal to` listOf(3)
-    }
-}
