+++ settings.gradle.kts
@@ -1,6 +1,7 @@
 rootProject.name = "dittnav-common-lib"
 include("dittnav-common-logging")
-include("dittnav-common-metrics")
-include("dittnav-common-security")
+include("dittnav-common-influx")
+include("dittnav-common-security-authenticated-user")
 include("dittnav-common-test")
 include("dittnav-common-utils")
+include("dittnav-common-evicting-cache")
